+++ rl_toolkit/policy/__init__.py
@@ -1,2 +1,3 @@
-from .off_policy import OffPolicy  # noqa
-from .sac import SAC  # noqa
+from .learner import Learner  # noqa
+from .agent import Agent  # noqa
+from .tester import Tester  # noqa