+++ rl_toolkit/policy/agent.py
@@ -1,16 +1,19 @@
-from abc import ABC, abstractmethod
+from rl_toolkit.networks import Actor
 
-import cv2
 import math
+import reverb
 import wandb
-import tensorflow as tf
+
 import numpy as np
+import tensorflow as tf
 
 
-class OffPolicy(ABC):
+class Agent:
     """
-    The base for Off-Policy algorithms
-    ==================================
+    Soft Actor-Critic
+    =================
+
+    Paper: https://arxiv.org/pdf/1812.05905.pdf
 
     Attributes:
         env: the instance of environment object
@@ -18,38 +21,35 @@ class OffPolicy(ABC):
         env_steps (int): maximum number of steps in each rollout
         gradient_steps (int): number of update steps after each rollout
         learning_starts (int): number of interactions before using policy network
-        buffer_capacity (int): the maximum size of experiences replay buffer
+        buffer_capacity (int): the capacity of experiences replay buffer
         batch_size (int): size of mini-batch used for training
+        actor_learning_rate (float): learning rate for actor's optimizer
+        critic_learning_rate (float): learning rate for critic's optimizer
+        alpha_learning_rate (float): learning rate for alpha's optimizer
         tau (float): the soft update coefficient for target networks
         gamma (float): the discount factor
+        model_a_path (str): path to the actor's model
+        model_c1_path (str): path to the critic_1's model
+        model_c2_path (str): path to the critic_2's model
         logging_wandb (bool): logging by WanDB
     """
 
     def __init__(
         self,
+        # ---
+        db_server: str,
+        # ---
         env,
-        # ---
         max_steps: int,
-        env_steps: int,
-        gradient_steps: int,
+        env_steps: int = 64,
+        learning_starts: int = 10000,
         # ---
-        learning_starts: int,
-        # ---
-        buffer_capacity: int,
-        batch_size: int,
-        # ---
-        tau: float,
-        gamma: float,
-        # ---
-        logging_wandb: bool,
+        logging_wandb: bool = False,
     ):
         self._env = env
         self._max_steps = max_steps
         self._env_steps = env_steps
-        self._gradient_steps = gradient_steps
         self._learning_starts = learning_starts
-        self._gamma = tf.constant(gamma)
-        self._tau = tf.constant(tau)
         self._logging_wandb = logging_wandb
 
         # check obseration's ranges
@@ -64,45 +64,62 @@ class OffPolicy(ABC):
 
             print("Observation cannot be normalized !\n")
 
-    @abstractmethod
-    def _get_action(self, state, deterministic):
-        ...
+        # Actor network (for agent)
+        self._actor = Actor(
+            state_shape=self._env.observation_space.shape,
+            action_shape=self._env.action_space.shape,
+        )
 
-    def _update_target(self, net, net_targ, tau):
-        for source_weight, target_weight in zip(
-            net.model.trainable_variables, net_targ.model.trainable_variables
-        ):
-            target_weight.assign(tau * source_weight + (1.0 - tau) * target_weight)
+        # prepare variable container
+        self._variables_container = {
+            "policy_variables": self._actor.model.variables,
+        }
 
-    @abstractmethod
-    def _update(self):
-        ...
+        # variables signature for variable container table
+        variable_container_signature = tf.nest.map_structure(
+            lambda variable: tf.TensorSpec(variable.shape, dtype=variable.dtype),
+            self._variables_container,
+        )
+        self._dtypes = tf.nest.map_structure(
+            lambda spec: spec.dtype, variable_container_signature
+        )
 
-    @abstractmethod
-    def save(self, path):
-        ...
+        # Initializes the reverb client
+        self.client = reverb.Client(f"{db_server}:8000")
+        self.tf_client = reverb.TFClient(server_address=f"{db_server}:8000")
 
-    @abstractmethod
-    def convert(self, path):
-        ...
+        # init Weights & Biases
+        if self._logging_wandb:
+            wandb.init(project="rl-toolkit")
 
-    @abstractmethod
-    def _logging_models(self):
-        ...
+            # Settings
+            wandb.config.max_steps = max_steps
+            wandb.config.env_steps = env_steps
+            wandb.config.learning_starts = learning_starts
 
     def _normalize_fn(self, obs):
-        # print(self._env.observation_space.low)
-        # print(self._env.observation_space.high)
-        # print(obs)
-
         # Min-max method
         obs = (obs - self._env.observation_space.low) / (
             self._env.observation_space.high - self._env.observation_space.low
         )
-
-        # print(obs)
-
         return obs
+
+    @tf.function
+    def _update_variables(self):
+        sample = self.tf_client.sample("variables", data_dtypes=[self._dtypes])
+        for variable, value in zip(
+            tf.nest.flatten(self._variables_container), tf.nest.flatten(sample.data[0])
+        ):
+            variable.assign(value)
+
+    @tf.function
+    def _get_action(self, state, deterministic):
+        a, _ = self._actor.predict(
+            tf.expand_dims(state, axis=0),
+            with_logprob=False,
+            deterministic=deterministic,
+        )
+        return tf.squeeze(a, axis=0)  # remove batch_size dim
 
     def _logging_train(self):
         print("=============================================")
@@ -112,7 +129,7 @@ class OffPolicy(ABC):
         print(f"TotalInteractions: {self._total_steps}")
         print("=============================================")
         print(
-            f"Training ... {math.floor(self._total_steps * 100.0 / self._max_steps)} %"
+            f"Playing ... {math.floor(self._total_steps * 100.0 / self._max_steps)} %"
         )
         if self._logging_wandb:
             wandb.log(
@@ -124,27 +141,7 @@ class OffPolicy(ABC):
                 step=self._total_steps,
             )
 
-    def _logging_test(self):
-        print("=============================================")
-        print(f"Epoch: {self._total_episodes}")
-        print(f"Score: {self._episode_reward}")
-        print(f"Steps: {self._episode_steps}")
-        print(f"TotalInteractions: {self._total_steps}")
-        print("=============================================")
-        print(
-            f"Testing ... {math.floor(self._total_steps * 100.0 / self._max_steps)} %"
-        )
-        if self._logging_wandb:
-            wandb.log(
-                {
-                    "epoch": self._total_episodes,
-                    "score": self._episode_reward,
-                    "steps": self._episode_steps,
-                },
-                step=self._total_steps,
-            )
-
-    def train(self):
+    def run(self):
         self._total_steps = 0
         self._total_episodes = 0
         self._episode_reward = 0.0
@@ -161,7 +158,7 @@ class OffPolicy(ABC):
                 self._update_variables()
 
                 # re-new noise matrix before every rollouts
-                self._actor_agent.reset_noise()
+                self._actor.reset_noise()
 
                 # collect rollouts
                 for _ in range(self._env_steps):
@@ -242,60 +239,3 @@ class OffPolicy(ABC):
 
                     # super critical !!!
                     self._last_obs = new_obs
-
-                # update models
-                if self._total_steps >= self._learning_starts:
-                    self._update()
-                    self._logging_models()
-
-    def test(self, render):
-        self._total_steps = 0
-        self._total_episodes = 0
-
-        # init video file
-        if render:
-            video_stream = cv2.VideoWriter(
-                "video/game.avi", cv2.VideoWriter_fourcc(*"MJPG"), 30, (640, 480)
-            )
-
-        # hlavny cyklus hry
-        while self._total_steps < self._max_steps:
-            self._episode_reward = 0.0
-            self._episode_steps = 0
-            done = False
-
-            self._last_obs = self._env.reset()
-            self._last_obs = self._normalize(self._last_obs)
-
-            # collect rollout
-            while not done:
-                # write to stream
-                if render:
-                    img_array = self._env.render(mode="rgb_array")
-                    img_array = cv2.resize(img_array, (640, 480))
-                    video_stream.write(img_array)
-
-                # Get the action
-                action = self._get_action(self._last_obs, deterministic=True).numpy()
-
-                # perform action
-                new_obs, reward, done, _ = self._env.step(action)
-                new_obs = self._normalize(new_obs)
-
-                # update variables
-                self._episode_reward += reward
-                self._episode_steps += 1
-                self._total_steps += 1
-
-                # super critical !!!
-                self._last_obs = new_obs
-
-            # increment episode
-            self._total_episodes += 1
-
-            # logovanie
-            self._logging_test()
-
-        # Release video file stream
-        if render:
-            video_stream.release()